Move test impl
